SPM-104 - fix merge issues
