+++ cli-node/lib/helpers.js
@@ -1,4 +1,3 @@
-const moment = require('moment-timezone');
 const util = require('util');
 const fs = require('fs');
 const marked = require('marked');
@@ -19,98 +18,26 @@ const {
 
 const {
   formatDateTime,
-  formatDateTimeWithComma,
-  formatTime,
-  calculateTimeDifference
+  calculateTimeDifference,
+  calculateTotalStake,
+  calculateTotalValue,
+  findFragments,
+  formatSchedules
 } = require('./formatters');
 
-const { infoSections } = require('./content');
+const { infoSections, tosAgreementContent } = require('./content');
 
 const TABLE_SIZE = 250;
 
-const calculateTotalValue = (stakeInfo, noUnassigned) => {
-  if (!stakeInfo || !stakeInfo.stake) return 0;
-
-  const dangling = stakeInfo.stake.dangling ? stakeInfo.stake.dangling : 0;
-  const unassigned =
-    !noUnassigned && stakeInfo.stake.unassigned
-      ? stakeInfo.stake.unassigned
-      : 0;
-
-  let totalPools = 0;
-
-  if (stakeInfo.stake.pools) {
-    stakeInfo.stake.pools.forEach(pool => {
-      totalPools += pool[1];
-    });
-  }
-
-  const total = totalPools + dangling + unassigned;
-  return total;
-};
-
-const calculateTotalStake = stakeInfo => calculateTotalValue(stakeInfo, true);
-
-const formatSchedules = schedules =>
-  schedules
-    .slice(0, schedules.length < TABLE_SIZE ? schedules.length : TABLE_SIZE)
-    .sort(
-      (a, b) =>
-        moment(b.scheduled_at_time).format('YYYYMMDDHHmmss') -
-        moment(a.scheduled_at_time).format('YYYYMMDDHHmmss')
-    )
-    .map(schedule => {
-      const scheduleDate = `${formatDateTimeWithComma(
-        schedule.scheduled_at_time
-      )} (${schedule.scheduled_at_date})`;
-
-      const startedAt = schedule.wake_at_time
-        ? formatTime(schedule.wake_at_time)
-        : 'TBD';
-
-      const finishedAt = schedule.finished_at_time
-        ? formatTime(schedule.finished_at_time)
-        : 'TBD';
-
-      return {
-        schedule: scheduleDate,
-        startedAt,
-        finishedAt
-      };
-    });
-
-const findFragments = (fragments, inputFragmentId) =>
-  fragments
-    .filter(fragment =>
-      inputFragmentId ? fragment.fragment_id === inputFragmentId : fragment
-    )
-    .slice(0, fragments.length < TABLE_SIZE ? fragments.length : TABLE_SIZE)
-    .sort(
-      (a, b) =>
-        moment(b.last_updated_at).format('YYYYMMDDHHmmss') -
-        moment(a.last_updated_at).format('YYYYMMDDHHmmss')
-    )
-    .map(fragment => {
-      const fragmentId = fragment.fragment_id;
-      const receivedAt = formatDateTimeWithComma(fragment.received_at);
-      const updatedAt = formatDateTimeWithComma(fragment.last_updated_at);
-      let { status } = fragment;
-
-      if (status && status.Rejected) {
-        status = `Rejected: ${status.Rejected.reason}`;
-      } else if (status && status.InABlock) {
-        status = `In a block: ${status.InABlock.date}`;
-      }
-
-      return {
-        fragmentId,
-        receivedAt,
-        updatedAt,
-        status
-      };
-    });
-
 const verifyConnection = nodeAddress => checkConnection(nodeAddress);
+
+const showTOSAgreement = callback => {
+  fs.readFile(tosAgreementContent, { encoding: 'utf8' }, (err, data) => {
+    if (err) return console.log(err);
+    console.log(marked(data));
+    if (callback) callback();
+  });
+};
 
 const showInfoSections = () => {
   console.log('spm 0.1.0');
@@ -222,7 +149,7 @@ const showStakeState = async nodeAddress => {
 
 const showLeaderSchedules = async nodeAddress => {
   const leaderSchedules = await getLeaderSchedules(nodeAddress);
-  const formattedSchedules = formatSchedules(leaderSchedules);
+  const formattedSchedules = formatSchedules(leaderSchedules, TABLE_SIZE);
   console.log('Leader Schedules: ');
   if (!formattedSchedules || formattedSchedules.length <= 0) {
     console.log('Leader schedules not found.');
@@ -233,7 +160,7 @@ const showLeaderSchedules = async nodeAddress => {
 
 const showFragmentLogs = async (nodeAddress, fragmentId) => {
   const fragmentLogs = await getFragmentLogs(nodeAddress);
-  const foundFragments = findFragments(fragmentLogs, fragmentId);
+  const foundFragments = findFragments(fragmentLogs, fragmentId, TABLE_SIZE);
   console.log('Fragment Logs: ');
   if (!foundFragments || foundFragments.length <= 0) {
     console.log('Fragment logs not found.');
@@ -255,5 +182,6 @@ module.exports = {
   showFragmentLogs,
   verifyConnection,
   showInfoContent,
-  showInfoSections
+  showInfoSections,
+  showTOSAgreement
 };
